Refactor: UserService내의 책임을 PhoneBookService로 분리
